pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dictation"
version = "0.1.0"
description = "Local speech-to-text dictation for Linux/Wayland using Whisper"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["dictation", "speech-to-text", "whisper", "voice", "transcription"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: GTK",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
dependencies = [
    "openai-whisper>=20231117",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "evdev>=1.6.0",
]

[project.optional-dependencies]
fast = ["faster-whisper>=1.0.0"]

[project.scripts]
dictation = "dictation.main:main"
dictation-settings = "dictation.ui.app:main"
dictation-doctor = "dictation.doctor:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
dictation = ["resources/*"]
